Add files via upload
The links between the pages can be found in the following places:
Homepage: Below the "More Info" Category
Page2: Directly below the title text
Page3: Below the first image of the page
ALSO, I'm really sorry for not styling this enough, I ran out of time because I was stupid and should've done this sooner. I'm really sorry.